requirements.txt
apache-airflow-providers-snowflake
snowflake-sqlalchemy
snowflake-connector-python
pytest
pytest-mock